增加聚合类型avg,min,max,sum实现
增加聚合类型avg,min,max,sum实现

+++ frostmourne-vue/src/views/alarm/list.vue
@@ -187,9 +187,15 @@ export default {
       })
     },
     remove(id) {
-      adminApi.delete(id).then(response => {
-        this.$message({ type: 'success', message: '删除成功', duration: 2000 })
-        this.fetchData()
+      this.$confirm('此操作将删除监控, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        adminApi.delete(id).then(response => {
+          this.$message({ type: 'success', message: '删除成功', duration: 2000 })
+          this.fetchData()
+        })
       })
     }
   }
